// src/formulas/formulas/contract/abstract/types/versionControl.ts
export type Namespace = string
export type ModuleVersion =
  | 'latest'
  | {
      version: string
    }
export type ModuleReference =
  | {
      account_base: number
    }
  | {
      native: Addr
    }
  | {
      adapter: Addr
    }
  | {
      app: number
    }
  | {
      standalone: number
    }
export type Addr = string
// ...
export interface ModuleInfo {
  name: string
  namespace: Namespace
  version: ModuleVersion
}
// ...
export interface Module {
  info: ModuleInfo
  reference: ModuleReference
}